+++ Changelog.md
@@ -4,6 +4,7 @@
 - Improved launch time ([#34](https://github.com/JoeSSS/calabash-launcher/pull/34))
 - Added "Reset to default" functionality ([#36](https://github.com/JoeSSS/calabash-launcher/pull/36))
 - Added possibility to choose path from drop-down menu
+- Added additional field in settings that allows to add run variables to the test run
 
 # 0.1