+++ Changelog.md
@@ -1,11 +1,23 @@
 # 0.2
 
-- Removed unused "Physical device" functionality ([#35](https://github.com/JoeSSS/calabash-launcher/pull/35))
-- Improved launch time ([#34](https://github.com/JoeSSS/calabash-launcher/pull/34))
+## Added
+
 - Added "Reset to default" functionality ([#36](https://github.com/JoeSSS/calabash-launcher/pull/36))
-- Added functionality to download APP from link
-- Transfered to new CommandsCore API
-- Added additional field in settings that allows to add run variables to the test run
+- Added functionality to download APP from link ([#69](https://github.com/xing/calabash-launcher/pull/69))
+- Added additional field in settings that allows to add run variables to the test run ([#66](https://github.com/xing/calabash-launcher/pull/66))
+
+## Removed
+
+- Removed the unused "Physical device" functionality ([#35](https://github.com/JoeSSS/calabash-launcher/pull/35))
+
+## Fixed
+
+- Fixed a crash when changing the Calabash folder path ([#71](https://github.com/xing/calabash-launcher/pull/71))
+
+## Miscellaneous
+
+- Updated to the new CommandsCore API ([#67](https://github.com/xing/calabash-launcher/pull/67))
+- Improved launch time ([#34](https://github.com/JoeSSS/calabash-launcher/pull/34))
 
 # 0.1