+++ Changelog.md
@@ -14,6 +14,7 @@
 ## Fixed
 
 - Fixed a crash when changing the Calabash folder path ([#71](https://github.com/xing/calabash-launcher/pull/71))
+- Screen synchronisation is terminating when the view controller changes ([#80](https://github.com/xing/calabash-launcher/pull/80/)) 
 
 ## Miscellaneous