+++ Changelog.md
@@ -6,6 +6,7 @@
 - Added functionality to download APP from link ([#69](https://github.com/xing/calabash-launcher/pull/69))
 - Added possibility to choose path from drop-down menu ([#62](https://github.com/xing/calabash-launcher/pull/62))
 - Added additional field in settings that allows to add run variables to the test run ([#66](https://github.com/xing/calabash-launcher/pull/66))
+- Added possibility to terminate test run ([#79](https://github.com/xing/calabash-launcher/pull/79))
 
 ## Removed
 
@@ -14,6 +15,8 @@
 ## Fixed
 
 - Fixed a crash when changing the Calabash folder path ([#71](https://github.com/xing/calabash-launcher/pull/71))
+- Fixed a crash when there are no simulators to select ([#82](https://github.com/xing/calabash-launcher/pull/82))
+- Screen synchronisation is terminating when the view controller changes ([#80](https://github.com/xing/calabash-launcher/pull/80/)) 
 
 ## Miscellaneous